pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.121.0",
    "groq>=0.33.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlparse>=0.5.3",
    "uvicorn>=0.38.0",
]
